Extract InstrumentingBackwardsCompatibilityVisitor
